requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
allennlp==2.9.3
allennlp-models==2.9.3
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
beautifulsoup4==4.11.1
black==22.6.0
blis==0.7.8
boto3==1.24.47
botocore==1.27.56
cached-path==1.1.5
cachetools==5.2.0
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.0.4
commonmark==0.9.1
conllu==4.4.1
cymem==2.0.6
datasets==2.4.0
dill==0.3.5.1
docker-pycreds==0.4.0
fairscale==0.4.6
filelock==3.6.0
flake8==5.0.4
flatdict==4.0.1
frozenlist==1.3.1
fsspec==2022.7.1
ftfy==6.1.1
gdown==4.5.1
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
h5py==3.7.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonnet==0.18.0
langcodes==3.3.0
lmdb==1.3.0
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
more-itertools==8.14.0
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.8
mypy-extensions==0.4.3
networkx==2.8.5
nltk==3.7
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.2
patsy==0.5.2
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
preshed==3.0.7
promise==2.3
protobuf==3.19.4
psutil==5.9.1
py==1.11.0
py-rouge==1.1
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pydantic==1.8.2
pyDeprecate==0.3.2
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pytokenizations==0.8.4
pytorch-lightning==1.7.0
pytz==2022.1
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.5.1
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.1.2
scipy==1.9.0
sentencepiece==0.1.97
sentry-sdk==1.9.5
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.2.4
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
statsmodels==0.13.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
termcolor==1.1.0
thinc==8.0.17
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
torchmetrics==0.8.0
tqdm==4.64.0
transformers==4.18.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
wandb==0.12.21
wasabi==0.10.1
wcwidth==0.2.5
Werkzeug==2.2.2
word2number==1.1
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
